temp_backend/requirements/dev.txt
# SHA1:0dd417a227b58e4cd8bdcf4ed8ea5988119e6fef
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
appdirs==1.4.4
    # via django-migration-linter
asttokens==2.4.1
    # via stack-data
comm==0.1.4
    # via ipykernel
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django-debug-toolbar==4.2.0
    # via -r parkour_app/requirements/dev.in
django-migration-linter==5.0.0
    # via -r parkour_app/requirements/dev.in
django-schema-viewer==0.4.0
    # via -r parkour_app/requirements/dev.in
executing==2.0.1
    # via stack-data
ipdb==0.13.13
    # via -r parkour_app/requirements/dev.in
ipykernel==6.26.0
    # via -r parkour_app/requirements/dev.in
ipython==8.17.2
    # via
    #   ipdb
    #   ipykernel
jedi==0.19.1
    # via ipython
jupyter-client==8.5.0
    # via ipykernel
jupyter-core==5.5.0
    # via
    #   ipykernel
    #   jupyter-client
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via werkzeug
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.5.8
    # via ipykernel
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
platformdirs==3.11.0
    # via jupyter-core
prompt-toolkit==3.0.39
    # via ipython
psutil==5.9.6
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.16.1
    # via
    #   ipython
    #   rich
pywatchman==1.4.1
    # via -r parkour_app/requirements/dev.in
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
rich==13.6.0
    # via -r parkour_app/requirements/dev.in
stack-data==0.6.3
    # via ipython
toml==0.10.2
    # via django-migration-linter
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.13.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
wcwidth==0.2.9
    # via prompt-toolkit
werkzeug==3.0.1
    # via -r parkour_app/requirements/dev.in
